Sonar fixes.
