+++ nanoFramework.Benchmark.Sample/Program.cs
@@ -4,7 +4,7 @@ using System.Threading;
 
 namespace nanoFramework.Benchmark.Sample
 {
-    public class Program
+    public static class Program
     {
         public static void Main()
         {
