+++ nanoFramework.Benchmark/Attributes/Abstract/LoggerAttribute.cs
@@ -3,9 +3,15 @@ using System;
 
 namespace nanoFramework.Benchmark.Attributes.Abstract
 {
+    /// <summary>
+    /// Base class for loggers.
+    /// </summary>
     [AttributeUsage(AttributeTargets.Class)]
     public abstract class LoggerAttribute : Attribute
     {
+        /// <summary>
+        /// Gets instance of logger.
+        /// </summary>
         public abstract ILogger Logger { get; }
     }
 }
