+++ nanoFramework.Benchmark/Attributes/Abstract/ParserAttribute.cs
@@ -3,10 +3,17 @@ using System;
 
 namespace nanoFramework.Benchmark.Attributes.Abstract
 {
+    /// <summary>
+    /// Base class for parser.
+    /// </summary>
     [AttributeUsage(AttributeTargets.Class, AllowMultiple = true)]
     public abstract class ParserAttribute : Attribute
     {
-        private static IResultParser parser;
+        private IResultParser parser;
+
+        /// <summary>
+        /// Gets instance of parser.
+        /// </summary>
         public IResultParser ResultParser
         {
             get
@@ -20,6 +27,9 @@ namespace nanoFramework.Benchmark.Attributes.Abstract
             }
         }
 
+        /// <summary>
+        /// Creates new instance of parser.
+        /// </summary>
         public abstract IResultParser CreateNewResultParser { get; }
     }
 }
