+++ nanoFramework.Benchmark/BenchmarkAttribute.cs
@@ -2,6 +2,9 @@
 
 namespace nanoFramework.Benchmark
 {
+    /// <summary>
+    /// Specify benchmark method.
+    /// </summary>
     [AttributeUsage(AttributeTargets.Method, AllowMultiple = false)]
     public sealed class BenchmarkAttribute : Attribute
     {
