+++ nanoFramework.Benchmark/Parser/ConsoleParser.cs
@@ -9,7 +9,7 @@ namespace nanoFramework.Benchmark.Parser
     internal class ConsoleParser : BaseTextParser, IResultParser
     {
         private const string ConsoleTableSeparator = "|";
-        private object lockObject = new();
+        private readonly object lockObject = new();
         private int[] columnsSize;
         public override void Parse(SingleBenchmarkResult benchmarkResult)
         {
@@ -22,7 +22,7 @@ namespace nanoFramework.Benchmark.Parser
 
         private int[] CalculateColumnSize(SingleBenchmarkResult benchmarkResult)
         {
-            var dataToDisplay = SingleBenchmarkResult.GetDataToDisplay();
+            var dataToDisplay = GetMethodsToDisplay();
             int[] columnSizes = new int[dataToDisplay.Length];
             for (int i = 0; i < dataToDisplay.Length; i++)
             {
