+++ nanoFramework.Benchmark/Parser/IResultParser.cs
@@ -3,6 +3,9 @@ using System;
 
 namespace nanoFramework.Benchmark.Parser
 {
+    /// <summary>
+    /// Base interface for reult parsers.
+    /// </summary>
     public interface IResultParser
     {
         void Parse(SingleBenchmarkResult benchmarkResult);
