+++ nanoFramework.Benchmark/Result/Attributes/DisplayAttribute.cs
@@ -3,7 +3,7 @@
 namespace nanoFramework.Benchmark.Result.Attributes
 {
     [AttributeUsage(AttributeTargets.Method, AllowMultiple = false)]
-    public sealed class DisplayAttribute : Attribute
+    sealed class DisplayAttribute : Attribute
     {
         public string Header { get; }
