+++ nanoFramework.Benchmark/SetupAttribute.cs
@@ -2,6 +2,9 @@
 
 namespace nanoFramework.Benchmark
 {
+    /// <summary>
+    /// Specifies a method that is invoked only once before running benchmark methods. 
+    /// </summary>
     [AttributeUsage(AttributeTargets.Method)]
     public class SetupAttribute : Attribute
     {
